# openapi/auto-gestao-api.yaml
openapi: 3.0.3
info:
  title: auto-gestao-api
  description: >
    Backend do domínio de condomínio (Condomínio, Endereço, Pessoa, Taxa Extra,
    Meio de Contato) do ecossistema Gestão Condial. Contrato contract-first:
    este arquivo é a fonte da verdade dos endpoints; as interfaces de
    controller e os modelos são gerados a partir dele pelo
    openapi-generator-maven-plugin.
  version: "1.0.0"
servers:
  - url: /
...
paths:
  /condominio:
    get:
      tags: [Condominio]
      operationId: listCondominios
      summary: Lista todos os condomínios cadastrados.
      responses:
        "200":
          description: Lista de condomínios.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Condominio"
    post:
      tags: [Condominio]
      operationId: createCondominio
      summary: Cria um novo condomínio.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Condominio"
      responses:
        "201":
          description: Condomínio criado.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Condominio"
        "400":
          description: Erro de validação dos parâmetros enviados.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListaDeErros"
        "404":
          description: Endereço informado não encontrado.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListaDeErros"
  /condominio/{idCondominio}:
    get:
      tags: [Condominio]
      operationId: findCondominio
      summary: Busca um condomínio pelo id.
      parameters:
        - name: idCondominio
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Condomínio encontrado.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Condominio"
    put:
      tags: [Condominio]
      operationId: updateCondominio
      summary: Atualiza um condomínio.
      description: Exige a permissão `CONDOMINIO_ESCREVER`.
      parameters:
        - name: idCondominio
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Condominio"
      responses:
        "200":
          description: Condomínio atualizado.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Condominio"
        "400":
          $ref: "#/components/responses/DadosInvalidos"
        "404":
          $ref: "#/components/responses/NaoEncontrado"
    delete:
      tags: [Condominio]
      operationId: deleteCondominio
      summary: Exclui um condomínio.
      description: >
        Exige a permissão `CONDOMINIO_ESCREVER`. Recusa com 409 se ainda houver unidades,
        moradores ou taxas vinculados — apagar em cascata levaria junto o cadastro de gente que
        mora lá.
      parameters:
        - name: idCondominio
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "204":
          description: Condomínio excluído.
        "404":
          $ref: "#/components/responses/NaoEncontrado"
        "409":
          $ref: "#/components/responses/EmUso"
  /condominio/taxaExtra:
    get:
      tags: [Condominio]
      operationId: listTaxasExtras
      summary: Lista as taxas extras.
      description: Exige a permissão `TAXA_LER`.
      parameters:
        - name: idCondominio
          in: query
          description: Restringe a um condomínio.
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Taxas extras.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TaxaExtra"
    post:
      tags: [Condominio]
      operationId: createTaxaExtra
      summary: Cria uma taxa extra para um condomínio.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaxaExtra"
      responses:
        "201":
          description: Taxa extra criada.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaxaExtra"
        "400":
          description: Erro de validação dos parâmetros enviados.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListaDeErros"
  /condominio/taxaExtra/{idTaxaExtra}:
    put:
      tags: [Condominio]
      operationId: updateTaxaExtra
      summary: Atualiza uma taxa extra.
      description: Exige a permissão `TAXA_ESCREVER`.
      parameters:
        - name: idTaxaExtra
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TaxaExtra"
      responses:
        "200":
          description: Taxa atualizada.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TaxaExtra"
        "400":
          $ref: "#/components/responses/DadosInvalidos"
        "404":
          $ref: "#/components/responses/NaoEncontrado"
    delete:
      tags: [Condominio]
      operationId: deleteTaxaExtra
      summary: Exclui uma taxa extra.
      description: >
        Exige a permissão `TAXA_ESCREVER`. Se for a última taxa do condomínio, o indicador
        `possuiTaxaExtra` volta para falso — do contrário o condomínio ficaria marcado como
        tendo taxa extra sem ter nenhuma.
      parameters:
        - name: idTaxaExtra
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "204":
          description: Taxa excluída.
        "404":
          $ref: "#/components/responses/NaoEncontrado"
  /condominio/taxaExtraPorCondominio/{idCondominio}:
    get:
      tags: [Condominio]
      operationId: findTaxaExtraCondominio
      summary: Lista as taxas extras de um condomínio.
      parameters:
        - name: idCondominio
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Lista de taxas extras do condomínio.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TaxaExtra"
...
  /unidade:
    get:
      tags: [Unidade]
      operationId: listUnidades
      summary: Lista as unidades.
      description: Exige a permissão `UNIDADE_LER`.
      parameters:
        - name: idCondominio
          in: query
          description: Restringe a um condomínio.
          schema:
            type: integer
            format: int64
        - name: ativa
          in: query
          schema:
            type: boolean
        - name: busca
          in: query
          description: Texto livre; casa com bloco e número.
          schema:
            type: string
      responses:
        "200":
          description: Unidades encontradas.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Unidade"
    post:
      tags: [Unidade]
      operationId: createUnidade
      summary: Cadastra uma unidade.
      description: >
        Exige a permissão `UNIDADE_ESCREVER`. A unidade nasce sem morador — vincular alguém é um
        passo separado, feito pelo cadastro do morador.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/Unidade"
      responses:
        "201":
          description: Unidade criada.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Unidade"
        "400":
          $ref: "#/components/responses/DadosInvalidos"
        "409":
          description: Já existe uma unidade com esse bloco e número no condomínio.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ListaDeErros"
...
components:
  responses:
    DadosInvalidos:
      description: Dados inválidos.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ListaDeErros"
    NaoEncontrado:
      description: Recurso não encontrado.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ListaDeErros"
    EmUso:
      description: O recurso ainda está vinculado a outros registros.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ListaDeErros"
  schemas:
    Unidade:
      type: object
      description: >
        Unidade autônoma de um condomínio. Existe independentemente de haver morador: o vínculo
        com uma Pessoa é opcional nos dois sentidos.

        Sem `required` de propósito: este mesmo schema aparece aninhado em `Pessoa.unidade`,
        onde serve só para referenciar a unidade pelo `id`. Exigir `numero` e `condominio` aqui
        rejeitaria `{"unidade": {"id": 7}}` na hora de vincular um morador. A obrigatoriedade
        vale na criação da unidade e é aplicada no `UnidadeDto`.
      properties:
        id:
          type: integer
          format: int64
          readOnly: true
        condominio:
          $ref: "#/components/schemas/Condominio"
        bloco:
          type: string
          maxLength: 30
          description: Opcional — condomínios de bloco único não usam.
        numero:
          type: string
          maxLength: 20
          description: Identificação da unidade. Texto, não número — "101", "A-3" e "Loja 2" valem.
        andar:
          type: integer
          format: int32
        fracaoIdeal:
          type: number
          format: double
          description: Participação da unidade no rateio das despesas.
        observacao:
          type: string
          maxLength: 300
        ativa:
          type: boolean
          default: true
        quantidadeDeMoradores:
          type: integer
          format: int32
          readOnly: true
          description: Quantos moradores estão vinculados. Zero é condição para poder excluir.
    Condominio:
      type: object
      properties:
        id:
          type: integer
          format: int64
          readOnly: true
        ddd:
          type: integer
          format: int32
        numeroTelefone:
          type: integer
          format: int64
        endereco:
          $ref: "#/components/schemas/Endereco"
        valorJuros:
          type: number
          format: double
        valorMulta:
          type: number
          format: double
        possuiTaxaExtra:
          type: boolean
          default: false
        valorTaxaCondominial:
          type: number
          format: double
        valorTaxasExtras:
          type: number
          format: double
          readOnly: true
          description: Somatório das taxas extras vigentes, calculado pelo servidor.
        valorTotalPorUnidade:
          type: number
          format: double
          readOnly: true
          description: Valor final por unidade (taxa condominial + taxas extras), calculado pelo servidor.
    Endereco:
      type: object
      properties:
        id:
          type: integer
          format: int64
          readOnly: true
        cep:
          type: integer
          format: int64
        endereco:
          type: string
        complemento:
          type: string
...
    TaxaExtra:
      type: object
      properties:
        id:
          type: integer
          format: int64
          readOnly: true
        condominio:
          $ref: "#/components/schemas/Condominio"
        valorTaxaExtra:
          type: number
          format: double
        numeroParcelas:
          type: integer
          format: int32
        descricaoTaxa:
          type: string
        dataCadastro:
          type: string
          format: date-time
    ErroDeParametro:
      type: object
      properties:
        parametro:
          type: string
        mensagem:
          type: string
    ListaDeErros:
      type: object
      properties:
        erros:
          type: array
          items:
            $ref: "#/components/schemas/ErroDeParametro"
        quantidadeDeErros:
          type: integer
          format: int32
          readOnly: true